+++ Source/ACE/Command/Handlers/AdminCommands.cs
@@ -8,6 +8,8 @@ using ACE.Network.GameMessages.Messages;
 using ACE.Common;
 using ACE.Database;
 using ACE.Network.Enum;
+using System.Collections.Generic;
+using log4net;
 
 namespace ACE.Command.Handlers
 {
@@ -87,15 +89,113 @@ namespace ACE.Command.Handlers
             // TODO: output
         }
 
-        // boot { account | char | iid } who
-        [CommandHandler("boot", AccessLevel.Sentinel, CommandHandlerFlag.RequiresWorld, 2)]
+        /// <summary>
+        /// Boots the Player or Account holder from the server and displays the CoC Violation Warning
+        /// </summary>
+        /// <remarks>
+        ///     TODO: 1. After the group messages are operational, Send out a message on the Audit Group Chat Channel and alert other admins of this command usage.
+        /// </remarks>
+        [CommandHandler("boot", AccessLevel.Sentinel, CommandHandlerFlag.None, 2, 
+            "Boots the Player or Account holder from the server and displays the CoC Violation Warning",
+            "{ account | char | iid } who")]
         public static void HandleBoot(Session session, params string[] parameters)
         {
-            // usage: @boot { account,char, iid} who
+            // usage: @boot { account, char, iid} who
             // This command boots the specified character out of the game.You can specify who to boot by account, character name, or player instance id.  'who' is the account / character / instance id to actually boot.
             // @boot - Boots the character out of the game.
+            // The first parameter may be only text
+            // The second paramater and proceeding parameters may also be a number or text, but the logic depends on the text from the first parameter.
+            AccountLookupType selectorType = AccountLookupType.Undef;
+            string bootName = "";
+            uint bootId = 0;
+            Session playerSession = null;
 
-            // TODO: output
+            // Loop through the AccountLookupEnum to attempt at matching the first parameter with a lookup type
+            foreach (string bootType in System.Enum.GetNames(typeof(AccountLookupType)))
+            {
+                // Check the FIRST character of the first parameter against the Enum dictionary
+                // If the first character matches (a,c,i) then the selector will be returned.
+                // This allows for users to type @boot char <name>, since char is a keyword in c#
+                // Switch case to Lower when matching
+                if (parameters[0].ToLower()[0] == bootType.ToLower()[0])
+                {
+                    // If found, selectorType will hold the correct AccoutLookupType
+                    // If this returns true, that means we were successful and can stop looping
+                    if (Enum.TryParse(bootType, out selectorType))
+                        break;
+                }
+            }
+
+            // Peform logic
+            if (selectorType != AccountLookupType.Undef)
+            {
+                // Extract the name from the parameters and get the name from the first parameter
+                if (selectorType == AccountLookupType.Account || selectorType == AccountLookupType.Character)
+                    bootName = Common.Extensions.CharacterNameExtensions.StringArrayToCharacterName(parameters, 1);
+
+                switch (selectorType)
+                {
+                    case AccountLookupType.Account:
+                        {
+                            playerSession = WorldManager.Find(bootName);
+                            if (playerSession != null)
+                                bootId = playerSession.Player.Guid.Low;
+                            break;
+                        }
+                    case AccountLookupType.Character:
+                        {
+                            playerSession = WorldManager.FindByPlayerName(bootName);
+                            if (playerSession != null)
+                                bootId = playerSession.Player.Guid.Low;
+                            break;
+                        }
+                    case AccountLookupType.Iid:
+                        {
+                            // Extract the Id from the parameters
+                            uint.TryParse(parameters[1], out bootId);
+                            playerSession = WorldManager.Find(new ObjectGuid(bootId));
+                            if (playerSession != null)
+                                bootName = playerSession.Player.Name;
+                            break;
+                        }
+                }
+
+                if (playerSession != null)
+                {
+                    string bootText = $"account or player: {bootName} id: {bootId}";
+                    // Boot the player
+                    playerSession.BootPlayer();
+
+                    // Send an update to the admin, but prevent sending too if the admin was the player being booted
+                    if (session != null)
+                    {
+                        // Log the player who initiated the boot
+                        bootText = "Player: " + session.Player.Name + " has booted " + bootText;
+                        if (session != playerSession)
+                            session.Network.EnqueueSend(new GameMessageSystemChat(bootText, ChatMessageType.Broadcast));
+                        // TODO: Send out a message on the Audit Group Chat Channel, to alert other admins of the boot
+                    }
+                    else
+                    {
+                        bootText = "Console booted " + bootText;
+                    }
+
+                    // log the boot to file
+                    ILog log = LogManager.GetLogger(System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);
+                    log.Info(bootText);
+
+                    // finish execution of command logic
+                    return;
+                }
+            } 
+
+            // Did not find a player
+            string errorText = $"Error locating the player or account to boot.";
+            // Send the error to a player or the console
+            if (session != null)
+                session.Network.EnqueueSend(new GameMessageSystemChat(errorText, ChatMessageType.Broadcast));
+            else
+                Console.WriteLine(errorText);
         }
 
         // cloak < on / off / player / creature >